refactor: extract connection pool to separate module
- Used node path alias to reference the module
- Used a connection pool instead of a single connection
