+++ src/db/test-db-connection.js
@@ -1,19 +1,12 @@
-import mysql from "mysql2/promise";
-import process from "node:process";
-
-const connectionSettings = {
-  host: process.env.DATABASE_HOST,
-  database: process.env.DATABASE_NAME,
-  user: process.env.DATABASE_USER,
-  password: process.env.DATABASE_PASSWORD,
-};
-const connection = await mysql.createConnection(connectionSettings);
+import { connectionPool } from "#src/db/pool.js";
 
 try {
+  const connection = await connectionPool.getConnection();
   const [results, fields] = await connection.query("select 1 + 1;");
   console.log({ results, fields });
+  connection.release();
 } catch (error) {
   console.error(error);
 }
 
-connection.end();
+connectionPool.end();
